+++ src/App.js
@@ -1,8 +1,6 @@
 import './App.css';
 import Navigation from './components/navigation/Navigation';
 import Display from './components/display/Display'
-import Navigation from './components/navigation/navigation';
-
 
 function App() {
   return (
