+++ src/components/about/About.js
@@ -1,0 +1,20 @@
+function About () {
+    return (
+        <div className='gitNoms'>
+            <h2>gitNoms</h2>
+            <p>
+            </p>
+        </div>
+        <div className='theCodingChefs'>
+            <h2>The Conding Chefs</h2>
+                <ul>
+                    <li>Steven Semco</li>
+                    <li>Woo-Sik Choi</li>
+                    <li>Collen Whitney</li>
+                    <li>Anselem Ogbennia</li>
+                </ul>
+        </div>
+    );
+}
+
+export default About;
