+++ src/components/display/Display.js
@@ -2,7 +2,7 @@ import { useState, useEffect } from "react";
 
 import Recipe from "./recipe/Recipe";
 
-const display = () => {
+const Display = () => {
     const API_URL = process.env.API_URL || 'http://localhost:4000/recipes';
     const [recipeData, setRecipeData] = useState([]);
 
@@ -33,4 +33,4 @@ const display = () => {
     );
 };
 
-export default display;
+export default Display;
